responsible ai/requirements.txt
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.20.0
matplotlib>=3.4.0
Pillow>=8.0.0
streamlit>=1.8.0
opencv-python>=4.5.0
h5py>=3.1.0
scikit-image>=0.18.0
tqdm>=4.60.0
